added final models. crud tests to be defined #20
